+++ NEWS.md
@@ -1,4 +1,5 @@
 * Add libraries script to templates
+* Add knitr options script to templates
 
 # clamour 0.0.0.9011 (2020-09-01)